requirements.txt
fastapi==0.134.0
uvicorn[standard]==0.41.0
google-cloud-storage==3.9.0
google-auth==2.48.0
google-api-python-client
aiohttp==3.13.3
pydantic==2.12.5
google-auth-oauthlib==1.3.0
gspread==6.0.0
google-cloud-aiplatform==1.78.0
pdfplumber==0.11.7
python-docx==1.1.2
openpyxl==3.1.5
python-dotenv==1.0.1
google-genai==1.18.0